pyproject.toml
[project]
name = "gonsters-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
packages = [{include = "app"}]
dependencies = [
    "fastapi-mqtt>=2.2.0",
    "fastapi[standard]>=0.123.9",
    "influxdb3-python>=0.16.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
    "pytest>=9.0.2",
    "pytest-cov>=4.1.0",
]


# Ruff configuration
[tool.ruff]
src = ["src"]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
# Enable pycodestyle (E), Pyflakes (F), isort (I), and more
select = ["E", "F", "I", "N", "W", "D", "UP", "B", "C4", "SIM", "ERA"]
ignore = ["E203", "E711", "E712", "E902"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
